Scene Modifications & Script Additions
Modified the bedroom scene by repositioning the plant, re-fixing player movement, and adjusting the collider on the bed to prevent camera clipping.

Modified the Party scene by adding the doors to each cell and adding a collider to the bar.

Added interactability to the doors and implemented a system requiring the player to face the object they wish to interact with.