Joe with Animations is in
Replaced old Joe with new Joe with eyes and animations, with all the functionality of Joe except he doesnt walk up to you if you approach him. Max idle time has also been increased to 60 for all npcs due to an imbalance in walking and idle times, and becuase problems arise with walking into things, so this could help in some capacity with the issue indirectly.